add new E_BECH32_NO_MEMORY error code to C API; clean up and better comment example programs
